Enhance to show both exceptions and timeouts as expected failures.
Previously exceptions and timeouts were not caught, so an `expectFail`
could not capture these situations and report them as expected
failures.

This change provides internal exception catching and (duplication of)
timing to allow `expectFail` to be used for those failures as well.

The ability to catch timeouts requires an unfortunate hack of slightly
shortening the timeout period, so that the wrapped timeout in
`expectFail` fires before the outer timeout in Tasty itself.  This is
an unfortunate kludge, but given the overall sequencing of Tasty and
the dis-inclination of the Tasty developers to integrate this package
into the core Tasty functionality, it's pretty much unavoidable.

Fixes Issue #6