Fix timeout handling for expected failures.
Normally, core Tasty provides timeout handling, but this is outside of
the local expected-failure wrapper, so a timeout shows up as an actual
failure, even if wrapped in an expected-failure wrapper.

This change gets the timeout value from the original test to apply
internally, and sets the Tasty timeout option for the test to
NoTimeout so that only local timeout management is occurring.